Optimised Entity serialisation for packet size.

+++ Shared/Include/Entities/Trait.hpp
@@ -5,7 +5,9 @@
 #ifndef R_TYPE_TRAITS_HPP
 #define R_TYPE_TRAITS_HPP
 
-enum Trait : int {
+#include <stdint.h>
+
+enum Trait : uint8_t {
     Drawable,
     Garbage
 };
